abstract usecases from the repo feature
