+++ app/src/main/java/com/hako/githubapi/di/AppModules.kt
@@ -2,8 +2,10 @@ package com.hako.githubapi.di
 
 import androidx.room.Room
 import com.hako.githubapi.data.database.GithubDatabase
-import com.hako.githubapi.data.retrofit.GithubClient
+import com.hako.githubapi.data.retrofit.github.GithubClient
 import com.hako.githubapi.data.retrofit.RemoteDatasource
+import com.hako.githubapi.domain.usecases.DeleteRepositories
+import com.hako.githubapi.domain.usecases.GetRepositories
 import com.hako.githubapi.features.pullRequest.PullListViewModel
 import com.hako.githubapi.features.repos.RepoListViewModel
 import org.koin.androidx.viewmodel.ext.koin.viewModel
@@ -20,11 +22,16 @@ val database = module {
     single { get<GithubDatabase>().repositoryDao() }
 }
 
-val views = module {
-    viewModel { RepoListViewModel(get()) }
-    viewModel { PullListViewModel() }
+val usecases = module {
+    single { GetRepositories(get()) }
+    single { DeleteRepositories(get()) }
 }
 
 val threads = module {
     single { Executors.newSingleThreadExecutor() }
+}
+
+val views = module {
+    viewModel { RepoListViewModel() }
+    viewModel { PullListViewModel() }
 }
